Move solidity.current.instructions and solidity.current.functionsByProgramCounter to solidity-utils
