+++ packages/debugger/package.json
@@ -20,7 +20,6 @@
     "test:debug": "NODE_ENV=testing node --inspect ./node_modules/.bin/mocha-webpack --webpack-config webpack/webpack.config-test.js --recursive"
   },
   "dependencies": {
-    "@truffle/code-utils": "^1.2.14",
     "@truffle/codec": "^0.4.2",
     "@truffle/compile-solidity": "^4.2.25",
     "@truffle/expect": "^0.0.13",
@@ -66,7 +65,6 @@
     "istanbul-instrumenter-loader": "^3.0.1",
     "mocha": "5.2.0",
     "mocha-webpack": "^1.1.0",
-    "node-interval-tree": "^1.3.3",
     "nyc": "^13.0.1",
     "remotedev-server": "^0.3.1",
     "webpack": "^3.8.1",
