+++ packages/solidity-utils/package.json
@@ -21,6 +21,10 @@
     "access": "public"
   },
   "dependencies": {
-    "debug": "^4.1.1"
+    "@truffle/code-utils": "^1.2.14",
+    "@truffle/codec": "^0.4.2",
+    "debug": "^4.1.1",
+    "json-pointer": "^0.6.0",
+    "node-interval-tree": "^1.3.3"
   }
 }
